requirements.txt
attrs==23.1.0
aws-cdk-lib==2.82.0
aws-cdk.asset-awscli-v1==2.2.185
aws-cdk.asset-kubectl-v20==2.1.1
aws-cdk.asset-node-proxy-agent-v5==2.0.155
cattrs==22.2.0
certifi==2023.5.7
constructs==10.2.41
datadog-api-client==2.12.0
exceptiongroup==1.1.1
importlib-resources==5.12.0
jsii==1.82.0
publication==0.0.3
python-dateutil==2.8.2
six==1.16.0
typeguard==2.13.3
typing_extensions==4.6.3
urllib3==2.0.2
zipp==3.15.0
